+++ test/fixtures/parser/bigquery/create_table_columns_partition_options.sql
@@ -1,0 +1,7 @@
+CREATE TABLE newtable (
+    x TIMESTAMP,
+    y INT64
+)
+PARTITION BY DATE(x)
+CLUSTER BY x, y
+OPTIONS(description="foo")